requirements.txt
Django==2.2.12
pycryptodome==3.9.7
